Factored new runner infrastructure into its own package
