+++ httomo/runner/gpu_utils.py
@@ -1,0 +1,9 @@
+from httomo.utils import gpu_enabled, xp
+
+
+def gpumem_cleanup():
+    """cleans up GPU memory and also the FFT plan cache"""
+    if gpu_enabled:
+        xp.get_default_memory_pool().free_all_blocks()
+        cache = xp.fft.config.get_plan_cache()
+        cache.clear()
